+++ profiles/core/inputs.tf
@@ -417,8 +417,9 @@ variable "cirrus_inputs" {
       critical = map(any)
     })
     process = object({
-      sqs_timeout           = number
-      sqs_max_receive_count = number
+      sqs_timeout                   = number
+      sqs_max_receive_count         = number
+      sqs_cross_account_sender_arns = optional(list(string))
     })
     state = object({
       timestream_magnetic_store_retention_period_in_days = number